requirements.txt
pynput==1.7.6
Pillow==10.2.0
firebase-admin==6.4.0
schedule==1.2.1
pywin32==306
PyInstaller==6.3.0
cryptography==42.0.2
